changepassword component added

+++ src/Components/accountSettings/changeNickname.jsx
@@ -5,7 +5,7 @@ import { bindActionCreators } from 'redux'
 
 import { userLogin } from '../../Store/actions.js'
 
-function ChangeNickname({JWTToken, userLogin, currentUserName}) {
+function ChangeNickname({ JWTToken, userLogin, currentUserName }) {
 
   const [newNickname, setNewNickname] = useState('')
   const [changingStatus, setChangingStatus] = useState({message: '', successful: null})
@@ -18,18 +18,27 @@ function ChangeNickname({JWTToken, userLogin, currentUserName}) {
 
   function changeNicknameHandler() {
 
+    const allowedSymbols = /^[A-Za-z0-9]+$/
+
     // some checks for new nickname
-    if (newNickname === '') {
-      setChangingStatus({
-        successful: false,
-        message: 'new nickname is empty'
-      })
-      return
-    }
     if (/\s/.test(newNickname)) {
       setChangingStatus({
         successful: false,
         message: 'nickname cannot contain whitespaces'
+      })
+      return
+    }
+    if (!allowedSymbols.test(newNickname)) {
+      setChangingStatus({
+        successful: false,
+        message: 'only letters and numbers allowed'
+      })
+      return
+    }
+    if (newNickname === '') {
+      setChangingStatus({
+        successful: false,
+        message: 'new nickname is empty'
       })
       return
     }
@@ -49,6 +58,15 @@ function ChangeNickname({JWTToken, userLogin, currentUserName}) {
     const config = {headers: {Authorization: 'Bearer ' + JWTToken}}
     axios.patch('/api/changeNickname', data, config)
       .then(res => {
+        
+        if (res.data.userExists) {
+          setChangingStatus({
+            successful: false,
+            message: 'this user exists yet'
+          })
+          return
+        }
+
         localStorage.setItem('JWTToken', res.data.newJWTToken)
         localStorage.setItem('LoginData', JSON.stringify({login:newNickname}))
         userLogin({username:newNickname, JWTToken: res.data.newJWTToken})
@@ -60,19 +78,19 @@ function ChangeNickname({JWTToken, userLogin, currentUserName}) {
       .catch(err => {
         setChangingStatus({
           successful: false,
-          message: 'this user exists yet'
+          message: 'something went wrong :('
         })
       })
   }
 
   return (
-    <div className='change-nickname' >
+    <div className='change-form' >
       <div>
         <input
           type='text'
           value={newNickname}
           onChange={event => setNewNickname(event.target.value)}
-          className='change-nickname-input'
+          className='change-form-input'
           placeholder='enter your new nickname'
         />
       </div>
